refactor: use constants file to keep constants and configs
