+++ .env.example
@@ -12,6 +12,7 @@ REACT_APP_API_BASE_URL=http://localhost/api
 
 # for Hardhat
 ETHEREUM_RPC_URL=https://mainnet.infura.io/v3/YOUR_KEY
+ADMIN_ADDRESS=0x0
 
 # for Prisma
 DATABASE_URL=postgres://username:password@localhost:5432/dbname
