+++ hardhat/scripts/deploy/00-mock-yearn/00-mock-yearn-deploy.ts
@@ -1,10 +1,14 @@
 import { DeployFunction } from 'hardhat-deploy/types';
 import { HardhatRuntimeEnvironment } from 'hardhat/types';
 
+import { TEST_ENV } from '../../../constants';
+
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
+  const useProxy = !hre.network.live;
+  if (TEST_ENV) return !useProxy;
+
   const { deployer } = await hre.getNamedAccounts();
   const { deploy } = hre.deployments;
-  const useProxy = !hre.network.live;
   const yRegistry = await deploy('MockRegistry', {
     contract: 'MockRegistry',
     from: deployer,
