+++ hardhat/scripts/deploy/00-mock-yearn/01-mock-yvusdc-deploy.ts
@@ -2,12 +2,15 @@ import { ethers } from 'hardhat';
 import { DeployFunction } from 'hardhat-deploy/types';
 import { HardhatRuntimeEnvironment } from 'hardhat/types';
 
+import { TEST_ENV } from '../../../constants';
 import { MockRegistry__factory } from '../../../typechain';
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
+  const useProxy = !hre.network.live;
+  if (TEST_ENV) return !useProxy;
+
   const { deployer } = await hre.getNamedAccounts();
   const { deploy } = hre.deployments;
-  const useProxy = !hre.network.live;
 
   const signer = await ethers.getSigner(deployer);
   const yRegistry = MockRegistry__factory.connect(
