+++ hardhat/test/00-ape-protocol/01-ape-vault-factory.ts
@@ -1,7 +1,7 @@
 import chai from 'chai';
 import { solidity } from 'ethereum-waffle';
 
-import { USDC_ADDRESS, USDC_YVAULT_ADDRESS } from '../constants';
+import { USDC_ADDRESS, USDC_YVAULT_ADDRESS } from '../../constants';
 import { createApeVault } from '../utils/ApeVault/createApeVault';
 import { DeploymentInfo, deployProtocolFixture } from '../utils/deployment';
 import { resetNetwork } from '../utils/network';
